Separate more generators into individual files
