+++ src/.srcfiles.yaml
@@ -89,8 +89,6 @@ Files:
     # the code to insert into the source and header files.
 
     generate/btn_widgets.cpp       # Button generation class
-    generate/checkbox_widgets.cpp  # wxCheckBox generation class
-    generate/combo_widgets.cpp     # Combo and choice generation classes
     generate/ctrl_widgets.cpp      # Calendar, GenericDir, Search Ctrl generation classes
     generate/dataview_widgets.cpp  # wxDataView generation classes
     generate/form_widgets.cpp      # Form generation classes
@@ -99,13 +97,11 @@ Files:
     generate/listbox_widgets.cpp   # ListBox generation classes
     generate/listctrl_widgets.cpp  # ListCtrl generation class
     generate/menu_widgets.cpp      # Menu generation classes
-    generate/misc_widgets.cpp      # Miscellaneous generation classes
     generate/panel_widgets.cpp     # Panel generation classes
     generate/picker_widgets.cpp    # Picker generation classes
     generate/project.cpp           # Project generator
     generate/radio_widgets.cpp     # Radio button and Radio box generation classes
     generate/ribbon_widgets.cpp    # Ribbon generation classes
-    generate/spin_widgets.cpp      # Spin and ScrollBar generation classes
     generate/styled_text.cpp       # wxStyledText (scintilla) generate
     generate/text_widgets.cpp      # Text generation classes
     generate/toolbar_widgets.cpp   # Toolbar generation classes
@@ -113,7 +109,29 @@ Files:
     generate/window_widgets.cpp    # Splitter and Scroll generation classes
     generate/wizard_form.cpp       # Wizard form class
 
-    generate/gen_static_box.cpp         # wxStaticBox generator
+    generate/gen_activity.cpp       # wxActivityIndicator generator
+    generate/gen_animation.cpp      # wxAnimationCtrl generator
+    generate/gen_banner_window.cpp  # wxBannerWindow generator
+    generate/gen_bitmap_combo.cpp   # wxBitmapComboBox generator
+    generate/gen_button.cpp         # wxButton generator
+    generate/gen_checkbox.cpp       # wxCheckBox generator
+    generate/gen_choice.cpp         # wxChoice generator
+    generate/gen_close_btn.cpp      # wxBitmapButton::CreateCloseButton generator
+    generate/gen_cmd_link_btn.cpp   # wxCommandLinkButton generator
+    generate/gen_combobox.cpp       # wxComboBox generator
+    generate/gen_custom_ctrl.cpp    # Custom Control generator
+    generate/gen_gauge.cpp          # wxGauge generator
+    generate/gen_hyperlink.cpp      # wxHyperlinkCtrl generator
+    generate/gen_infobar.cpp        # wxInfoBar generator
+    generate/gen_scrollbar.cpp      # wxScrollBar generator
+    generate/gen_slider.cpp         # wxSlider generator
+    generate/gen_spin_btn.cpp       # wxSpinButton generator
+    generate/gen_spin_ctrl.cpp      # wxSpinButton generator
+    generate/gen_static_bmp.cpp     # wxStaticBitmap generator
+    generate/gen_static_box.cpp     # wxStaticBox generator
+    generate/gen_static_line.cpp    # wxStaticLine generator
+    generate/gen_status_bar.cpp     # wxStatusBar generator
+    generate/gen_toggle_btn.cpp     # wxToggleButton generator
 
     generate/gen_box_sizer.cpp          # wxBoxSizer generator
     generate/gen_flexgrid_sizer.cpp     # wxFlexGridSizer generator
