+++ src/generate/gen_choice.cpp
@@ -1,0 +1,204 @@
+//////////////////////////////////////////////////////////////////////////
+// Purpose:   wxChoice generator
+// Author:    Ralph Walden
+// Copyright: Copyright (c) 2020-2022 KeyWorks Software (Ralph Walden)
+// License:   Apache License -- see ../../LICENSE
+/////////////////////////////////////////////////////////////////////////////
+
+#include <wx/choice.h>  // wxChoice class interface
+
+#include "gen_common.h"     // GeneratorLibrary -- Generator classes
+#include "gen_xrc_utils.h"  // Common XRC generating functions
+#include "node.h"           // Node class
+#include "pugixml.hpp"      // xml read/write/create/process
+#include "utils.h"          // Utility functions that work with properties
+
+#include "gen_choice.h"
+
+//////////////////////////////////////////  ChoiceGenerator  //////////////////////////////////////////
+
+wxObject* ChoiceGenerator::CreateMockup(Node* node, wxObject* parent)
+{
+    auto widget = new wxChoice(wxStaticCast(parent, wxWindow), wxID_ANY, DlgPoint(parent, node, prop_pos),
+                               DlgSize(parent, node, prop_size), 0, nullptr, GetStyleInt(node));
+
+    if (node->HasValue(prop_contents))
+    {
+        auto array = ConvertToArrayString(node->prop_as_string(prop_contents));
+        for (auto& iter: array)
+            widget->Append(iter.wx_str());
+
+        if (node->HasValue(prop_selection_string))
+        {
+            widget->SetStringSelection(node->prop_as_wxString(prop_selection_string));
+        }
+        else
+        {
+            int sel = node->prop_as_int(prop_selection_int);
+            if (sel > -1 && sel < (to_int) array.size())
+                widget->SetSelection(sel);
+        }
+    }
+
+    widget->Bind(wxEVT_LEFT_DOWN, &BaseGenerator::OnLeftClick, this);
+
+    return widget;
+}
+
+bool ChoiceGenerator::OnPropertyChange(wxObject* widget, Node* node, NodeProperty* prop)
+{
+    if (!node->HasValue(prop_contents))
+        return false;
+
+    if (prop->isProp(prop_selection_string))
+    {
+        wxStaticCast(widget, wxChoice)->SetStringSelection(prop->as_wxString());
+        return true;
+    }
+    else if (prop->isProp(prop_selection_int))
+    {
+        wxStaticCast(widget, wxChoice)->SetSelection(prop->as_int());
+        return true;
+    }
+    return false;
+}
+
+std::optional<ttlib::cstr> ChoiceGenerator::GenConstruction(Node* node)
+{
+    ttlib::cstr code;
+    if (node->IsLocal())
+        code << "auto ";
+    code << node->get_node_name() << GenerateNewAssignment(node);
+    code << GetParentName(node) << ", " << node->prop_as_string(prop_id);
+
+    if (node->prop_as_string(prop_window_name).empty() && node->prop_as_string(prop_style).empty() &&
+        node->prop_as_string(prop_window_style).empty())
+    {
+        GeneratePosSizeFlags(node, code);
+    }
+    else
+    {
+        // We have to generate a default validator before the window name, which GeneratePosSizeFlags doesn't do. We don't
+        // actually need that validator, since GenSettings will create it, but we have to supply something before the window
+        // name.
+
+        code << ", ";
+        GenPos(node, code);
+        code << ", ";
+        GenSize(node, code);
+        code << ", 0, nullptr, ";
+        GenStyle(node, code);
+        if (node->HasValue(prop_window_name))
+        {
+            code << ", wxDefaultValidator, " << node->prop_as_string(prop_window_name);
+        }
+        code << ");";
+    }
+
+    return code;
+}
+
+std::optional<ttlib::cstr> ChoiceGenerator::GenSettings(Node* node, size_t& /* auto_indent */)
+{
+    ttlib::cstr code;
+
+    if (node->prop_as_bool(prop_focus))
+    {
+        if (code.size())
+            code << '\n';
+        code << node->get_node_name() << "->SetFocus()";
+    }
+
+    if (node->HasValue(prop_contents))
+    {
+        auto array = ConvertToArrayString(node->prop_as_string(prop_contents));
+        for (auto& iter: array)
+        {
+            if (code.size())
+                code << "\n";
+            code << node->get_node_name() << "->Append(" << GenerateQuotedString(iter) << ");";
+        }
+
+        if (node->HasValue(prop_selection_string))
+        {
+            code << "\n";
+            if (node->HasValue(prop_validator_variable))
+            {
+                code << node->prop_as_string(prop_validator_variable) << " = ";
+                code << GenerateQuotedString(node, prop_selection_string) << ";  // set validator variable";
+            }
+            else
+            {
+                code << node->get_node_name() << "->SetStringSelection(";
+                code << GenerateQuotedString(node, prop_selection_string) << ");";
+            }
+        }
+        else
+        {
+            int sel = node->prop_as_int(prop_selection_int);
+            if (sel > -1 && sel < (to_int) array.size())
+            {
+                code << "\n";
+                code << node->get_node_name() << "->SetSelection(" << node->prop_as_string(prop_selection_int) << ");";
+            }
+        }
+    }
+
+    return code;
+}
+
+std::optional<ttlib::cstr> ChoiceGenerator::GenEvents(NodeEvent* event, const std::string& class_name)
+{
+    return GenEventCode(event, class_name);
+}
+
+bool ChoiceGenerator::GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr)
+{
+    InsertGeneratorInclude(node, "#include <wx/choice.h>", set_src, set_hdr);
+    return true;
+}
+
+int ChoiceGenerator::GenXrcObject(Node* node, pugi::xml_node& object, bool add_comments)
+{
+    auto result = node->GetParent()->IsSizer() ? BaseGenerator::xrc_sizer_item_created : BaseGenerator::xrc_updated;
+    auto item = InitializeXrcObject(node, object);
+
+    GenXrcObjectAttributes(node, item, "wxChoice");
+
+    if (node->HasValue(prop_contents))
+    {
+        auto content = item.append_child("content");
+        auto array = ConvertToArrayString(node->prop_as_string(prop_contents));
+        for (auto& iter: array)
+        {
+            content.append_child("item").text().set(iter);
+        }
+    }
+
+    if (node->HasValue(prop_selection_string))
+        item.append_child("value").text().set(node->prop_as_string(prop_selection_string));
+
+    // Older versions of wxWidgets didn't support setting the selection via the value property,
+    // so we add the property here even if the above is set.
+    if (node->prop_as_int(prop_selection_int) >= 0)
+        item.append_child("selection").text().set(node->prop_as_string(prop_selection_int));
+
+    GenXrcStylePosSize(node, item);
+    GenXrcWindowSettings(node, item);
+
+    if (add_comments)
+    {
+        if (node->HasValue(prop_selection_string))
+        {
+            ADD_ITEM_COMMENT("You cannot use selection_string for the selection in XRC.")
+        }
+        GenXrcComments(node, item);
+    }
+
+    return result;
+}
+
+void ChoiceGenerator::RequiredHandlers(Node* /* node */, std::set<std::string>& handlers)
+{
+    handlers.emplace("wxChoiceXmlHandler");
+}
