+++ src/generate/gen_combobox.cpp
@@ -1,0 +1,217 @@
+/////////////////////////////////////////////////////////////////////////////
+// Purpose:   wxComboBox generator
+// Author:    Ralph Walden
+// Copyright: Copyright (c) 2020-2022 KeyWorks Software (Ralph Walden)
+// License:   Apache License -- see ../../LICENSE
+/////////////////////////////////////////////////////////////////////////////
+
+#include <wx/combobox.h>  // wxComboBox declaration
+
+#include "gen_common.h"  // GeneratorLibrary -- Generator classes
+#include "node.h"        // Node class
+#include "utils.h"       // Utility functions that work with properties
+
+#include "gen_combobox.h"
+
+wxObject* ComboBoxGenerator::CreateMockup(Node* node, wxObject* parent)
+{
+    auto widget = new wxComboBox(wxStaticCast(parent, wxWindow), wxID_ANY, wxEmptyString, DlgPoint(parent, node, prop_pos),
+                                 DlgSize(parent, node, prop_size), 0, nullptr, GetStyleInt(node));
+
+    if (node->HasValue(prop_hint) && !node->prop_as_string(prop_style).contains("wxCB_READONLY"))
+        widget->SetHint(node->prop_as_wxString(prop_hint));
+
+    if (node->HasValue(prop_contents))
+    {
+        auto array = ConvertToArrayString(node->prop_as_string(prop_contents));
+        for (auto& iter: array)
+            widget->Append(iter.wx_str());
+
+        if (node->HasValue(prop_selection_string))
+        {
+            widget->SetStringSelection(node->prop_as_wxString(prop_selection_string));
+        }
+        else
+        {
+            int sel = node->prop_as_int(prop_selection_int);
+            if (sel > -1 && sel < (to_int) array.size())
+                widget->SetSelection(sel);
+        }
+    }
+
+    // Note that this event only gets fired of the drop-down button is clicked. Clicking in the edit area does not
+    // generate an event (or at least it doesn't on Windows 10).
+
+    widget->Bind(wxEVT_LEFT_DOWN, &BaseGenerator::OnLeftClick, this);
+
+    return widget;
+}
+
+bool ComboBoxGenerator::OnPropertyChange(wxObject* widget, Node* node, NodeProperty* prop)
+{
+    if (!node->HasValue(prop_contents))
+        return false;
+
+    if (prop->isProp(prop_selection_string))
+    {
+        wxStaticCast(widget, wxComboBox)->SetStringSelection(prop->as_wxString());
+        return true;
+    }
+    else if (prop->isProp(prop_selection_int))
+    {
+        wxStaticCast(widget, wxComboBox)->SetSelection(prop->as_int());
+        return true;
+    }
+    return false;
+}
+
+std::optional<ttlib::cstr> ComboBoxGenerator::GenConstruction(Node* node)
+{
+    ttlib::cstr code;
+    if (node->IsLocal())
+        code << "auto ";
+    code << node->get_node_name() << GenerateNewAssignment(node);
+    code << GetParentName(node) << ", " << node->prop_as_string(prop_id) << ", ";
+
+    // We don't add any strings until after the constructor, so if an initial selection string is specified, we set it
+    // after all strings have been appended.
+    code << "wxEmptyString";
+
+    if (node->prop_as_string(prop_window_name).empty() && node->prop_as_string(prop_style).empty() &&
+        node->prop_as_string(prop_window_style).empty())
+    {
+        GeneratePosSizeFlags(node, code);
+    }
+    else
+    {
+        // We have to generate a default validator before the window name, which GeneratePosSizeFlags doesn't do. We don't
+        // actually need that validator, since GenSettings will create it, but we have to supply something before the window
+        // name.
+
+        code << ", ";
+        GenPos(node, code);
+        code << ", ";
+        GenSize(node, code);
+        code << ", 0, nullptr, ";
+        GenStyle(node, code);
+        if (node->HasValue(prop_window_name))
+        {
+            code << ", wxDefaultValidator, " << node->prop_as_string(prop_window_name);
+        }
+        code << ");";
+    }
+
+    return code;
+}
+
+std::optional<ttlib::cstr> ComboBoxGenerator::GenSettings(Node* node, size_t& /* auto_indent */)
+{
+    ttlib::cstr code;
+
+    if (node->HasValue(prop_hint) && !node->prop_as_string(prop_style).contains("wxCB_READONLY"))
+    {
+        if (code.size())
+            code << '\n';
+        code << node->get_node_name() << "->SetHint(" << GenerateQuotedString(node->prop_as_string(prop_hint)) << ");";
+    }
+
+    if (node->prop_as_bool(prop_focus))
+    {
+        if (code.size())
+            code << '\n';
+        code << node->get_node_name() << "->SetFocus()";
+    }
+
+    if (node->HasValue(prop_contents))
+    {
+        auto array = ConvertToArrayString(node->prop_as_string(prop_contents));
+        for (auto& iter: array)
+        {
+            if (code.size())
+                code << "\n";
+            code << node->get_node_name() << "->Append(" << GenerateQuotedString(iter) << ");";
+        }
+
+        if (node->HasValue(prop_selection_string))
+        {
+            code << "\n";
+            if (node->HasValue(prop_validator_variable))
+            {
+                code << node->prop_as_string(prop_validator_variable) << " = ";
+                code << GenerateQuotedString(node, prop_selection_string) << ";  // set validator variable";
+            }
+            else
+            {
+                code << node->get_node_name() << "->SetStringSelection(";
+                code << GenerateQuotedString(node, prop_selection_string) << ");";
+            }
+        }
+        else
+        {
+            int sel = node->prop_as_int(prop_selection_int);
+            if (sel > -1 && sel < (to_int) array.size())
+            {
+                code << "\n";
+                code << node->get_node_name() << "->SetSelection(" << node->prop_as_string(prop_selection_int) << ");";
+            }
+        }
+    }
+
+    return code;
+}
+
+std::optional<ttlib::cstr> ComboBoxGenerator::GenEvents(NodeEvent* event, const std::string& class_name)
+{
+    return GenEventCode(event, class_name);
+}
+
+bool ComboBoxGenerator::GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr)
+{
+    InsertGeneratorInclude(node, "#include <wx/combobox.h>", set_src, set_hdr);
+    return true;
+}
+
+int ComboBoxGenerator::GenXrcObject(Node* node, pugi::xml_node& object, bool add_comments)
+{
+    auto result = node->GetParent()->IsSizer() ? BaseGenerator::xrc_sizer_item_created : BaseGenerator::xrc_updated;
+    auto item = InitializeXrcObject(node, object);
+
+    GenXrcObjectAttributes(node, item, "wxComboBox");
+
+    if (node->HasValue(prop_contents))
+    {
+        auto content = item.append_child("content");
+        auto array = ConvertToArrayString(node->prop_as_string(prop_contents));
+        for (auto& iter: array)
+        {
+            content.append_child("item").text().set(iter);
+        }
+    }
+
+    if (node->HasValue(prop_selection_string))
+        item.append_child("value").text().set(node->prop_as_string(prop_selection_string));
+    else if (node->prop_as_int(prop_selection_int) >= 0)
+        item.append_child("selection").text().set(node->prop_as_string(prop_selection_int));
+
+    if (node->HasValue(prop_hint) && !node->prop_as_string(prop_style).contains("wxCB_READONLY"))
+        item.append_child("hint").text().set(node->prop_as_string(prop_hint));
+
+    GenXrcStylePosSize(node, item);
+    GenXrcWindowSettings(node, item);
+
+    if (add_comments)
+    {
+        if (node->HasValue(prop_selection_string))
+        {
+            ADD_ITEM_COMMENT("You cannot use selection_string for the selection in XRC.")
+        }
+        GenXrcComments(node, item);
+    }
+
+    return result;
+}
+
+void ComboBoxGenerator::RequiredHandlers(Node* /* node */, std::set<std::string>& handlers)
+{
+    handlers.emplace("wxComboBoxXmlHandler");
+}
