+++ src/generate/gen_spin_ctrl.h
@@ -1,5 +1,5 @@
 /////////////////////////////////////////////////////////////////////////////
-// Purpose:   Combo and choice component classes
+// Purpose:   Spin and ScrollBar component classes
 // Author:    Ralph Walden
 // Copyright: Copyright (c) 2020-2021 KeyWorks Software (Ralph Walden)
 // License:   Apache License -- see ../../LICENSE
@@ -9,7 +9,7 @@
 
 #include "base_generator.h"  // BaseGenerator -- Base Generator class
 
-class ComboBoxGenerator : public BaseGenerator
+class SpinCtrlGenerator : public BaseGenerator
 {
 public:
     wxObject* CreateMockup(Node* node, wxObject* parent) override;
@@ -25,31 +25,19 @@ public:
     void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
 };
 
-class ChoiceGenerator : public BaseGenerator
+class SpinCtrlDoubleGenerator : public BaseGenerator
 {
 public:
     wxObject* CreateMockup(Node* node, wxObject* parent) override;
-    bool OnPropertyChange(wxObject* widget, Node* node, NodeProperty* prop) override;
 
     std::optional<ttlib::cstr> GenConstruction(Node* node) override;
     std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
     std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
+
+    bool AllowPropertyChange(wxPropertyGridEvent*, NodeProperty*, Node*) override;
 
     bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
 
     int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
     void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class BitmapComboBoxGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-    bool OnPropertyChange(wxObject* widget, Node* node, NodeProperty* prop) override;
-
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
-
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
 };
