+++ src/generate/gen_static_bmp.h
@@ -1,0 +1,25 @@
+//////////////////////////////////////////////////////////////////////////
+// Purpose:   wxStaticBitmap generator
+// Author:    Ralph Walden
+// Copyright: Copyright (c) 2020-2022 KeyWorks Software (Ralph Walden)
+// License:   Apache License -- see ../../LICENSE
+/////////////////////////////////////////////////////////////////////////////
+
+#pragma once
+
+#include "base_generator.h"  // BaseGenerator -- Base Generator class
+
+class StaticBitmapGenerator : public BaseGenerator
+{
+public:
+    wxObject* CreateMockup(Node* node, wxObject* parent) override;
+
+    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
+    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
+    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
+
+    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
+
+    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
+    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
+};
