--- src/generate/misc_widgets.h
@@ -1,173 +1,0 @@
-/////////////////////////////////////////////////////////////////////////////
-// Purpose:   Miscellaneous component classes
-// Author:    Ralph Walden
-// Copyright: Copyright (c) 2020-2021 KeyWorks Software (Ralph Walden)
-// License:   Apache License -- see ../../LICENSE
-/////////////////////////////////////////////////////////////////////////////
-
-#pragma once
-
-#include <wx/infobar.h>  // declaration of wxInfoBarBase defining common API of wxInfoBar
-#include <wx/timer.h>    // wxTimer, wxStopWatch and global time-related functions
-
-#include "base_generator.h"  // BaseGenerator -- Base Generator class
-
-class ActivityIndicatorGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class AnimationGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class BannerWindowGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class CustomControl : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-};
-
-class StaticLineGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class StatusBarGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
-
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class StaticBitmapGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
-
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class GaugeGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-    bool OnPropertyChange(wxObject* widget, Node* node, NodeProperty* prop) override;
-
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
-
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class SliderGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-    bool OnPropertyChange(wxObject* widget, Node* node, NodeProperty* prop) override;
-
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
-
-    bool AllowPropertyChange(wxPropertyGridEvent*, NodeProperty*, Node*) override;
-
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class HyperlinkGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
-    bool IsGeneric(Node*) override;
-
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class InfoBarGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-
-protected:
-    void OnButton(wxCommandEvent& event);
-    void OnTimer(wxTimerEvent&);
-
-private:
-    wxTimer m_timer;
-    wxInfoBar* m_infobar;
-};
