--- src/generate/spin_widgets.h
@@ -1,72 +1,0 @@
-/////////////////////////////////////////////////////////////////////////////
-// Purpose:   Spin and ScrollBar component classes
-// Author:    Ralph Walden
-// Copyright: Copyright (c) 2020-2021 KeyWorks Software (Ralph Walden)
-// License:   Apache License -- see ../../LICENSE
-/////////////////////////////////////////////////////////////////////////////
-
-#pragma once
-
-#include "base_generator.h"  // BaseGenerator -- Base Generator class
-
-class SpinCtrlGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-    bool OnPropertyChange(wxObject* widget, Node* node, NodeProperty* prop) override;
-
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
-
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class SpinCtrlDoubleGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
-
-    bool AllowPropertyChange(wxPropertyGridEvent*, NodeProperty*, Node*) override;
-
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class SpinButtonGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
-    std::optional<ttlib::cstr> GenSettings(Node* node, size_t& auto_indent) override;
-
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
-
-class ScrollBarGenerator : public BaseGenerator
-{
-public:
-    wxObject* CreateMockup(Node* node, wxObject* parent) override;
-
-    std::optional<ttlib::cstr> GenConstruction(Node* node) override;
-    std::optional<ttlib::cstr> GenEvents(NodeEvent* event, const std::string& class_name) override;
-
-    bool GetIncludes(Node* node, std::set<std::string>& set_src, std::set<std::string>& set_hdr) override;
-
-    int GenXrcObject(Node*, pugi::xml_node& /* object */, bool /* add_comments */) override;
-    void RequiredHandlers(Node*, std::set<std::string>& /* handlers */) override;
-};
